requirements.txt
google-auth>=2.0.0
google-auth-oauthlib>=1.0.0
google-api-python-client>=2.0.0
pyyaml>=6.0
playwright>=1.40.0
